Make generated Radicale rights file compatible with Radicale web interface
Using the Radicale web interface requires non-calendar/addressbook read access (`permissions` value `R`) to the root collection for all rules that apply. Adding this as a separate rule ([as suggested in the Radicale documentation](https://radicale.org/v3.html#authentication-and-rights)) in front of all other rules does not work, for some reason the same rule needs to match both the user’s principal collection and the root collection or the web interface fails with a permission error.